feat(common): fix some stuff, need to continue refactoring singleton_config for display
