+++ client/src/screen/texture_config_helper.rs
